Move StatusBox component to console shared package
