+++ frontend/packages/console-shared/locales/en/console-shared.json
@@ -235,6 +235,9 @@
   "Click": "Click",
   "Right click": "Right click",
   "Drag + Drop": "Drag + Drop",
+  "404: Not Found": "404: Not Found",
+  "{{labels}} content is not available in the catalog at this time due to loading failures.": "{{labels}} content is not available in the catalog at this time due to loading failures.",
+  "Timed out fetching new data. The data below is stale.": "Timed out fetching new data. The data below is stale.",
   "Information": "Information",
   "In progress": "In progress",
   "Healthy": "Healthy",
