+++ frontend/packages/console-shared/src/components/status/index.tsx
@@ -1,6 +1,7 @@
 export * from './icons';
 export * from './statuses';
 export * from './Status';
+export * from './StatusBox';
 
 export { default as SecondaryStatus } from './SecondaryStatus';
 export { default as LinkStatus } from './LinkStatus';
