+++ frontend/public/components/utils/status-box.ts
@@ -1,4 +1,3 @@
-export * from './StatusBox';
-
+export * from '@console/shared/src/components/status/StatusBox';
 export * from '@console/shared/src/components/loading';
 export * from '@console/shared/src/components/empty-state';
